Refactor date functionality into separate util.
